requirements.txt
google-generativeai==0.3.0
gTTS==2.4.0
moviepy==1.0.3
python-dotenv==1.0.0
